requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Naked==0.1.31
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
